Code Gulpified
Code broken down into managable modules, used Gulp to compile all the typescript files into bundle.js.
