+++ 7-Logical-Agents/index.html
@@ -3,7 +3,7 @@
 <head>
     <title>7 Logical Agents </title>
     <link rel="stylesheet" href="../styles.css">
-    <script src="jquery.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
     <script type="text/javascript" src="../main.js"></script>
 </head>
 <body>
@@ -64,8 +64,8 @@
 
         </div>
     </div>
-    <script src="svg.min.js"></script>
-    <script src="grid.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/svg.js/2.7.1/svg.min.js"></script>
+    <script src="bundle.js"></script>
     <script> $(function() { $('header').load("../header.html"); }); </script>
 </body>
 </html>
