+++ 7-Logical-Agents/main.ts
@@ -1,0 +1,9 @@
+import { Grid } from "./grid";
+
+const game: Grid = new Grid();
+game.getTile(3, 1).hasPit = true;
+game.getTile(3, 3).hasPit = true;
+game.getTile(4, 4).hasPit = true;
+game.getTile(1, 3).hasWumpus = true;
+game.getTile(2, 3).hasGold = true;
+game.sensorUpdate();
