+++ 7-Logical-Agents/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "files": [
+    "grid.ts"
+  ],
+  "compilerOptions": {
+    "target": "es5",                          /* Specify ECMAScript target version: 'ES5'. */
+    "module": "commonjs",                     /* Specify module code generation: 'commonjs' */
+    "strict": true,                           /* Enable all strict type-checking options. */
+    "esModuleInterop": true                   /* Enables emit interoperability between CommonJS and ES Modules */
+  }
+}
