+++ Gulpfile.babel.js
@@ -1,6 +1,8 @@
 import gulp from 'gulp';
 import ghPages from 'gulp-gh-pages';
-import jspm from 'jspm';
+let browserify = require("browserify");
+let source = require('vinyl-source-stream');
+let tsify = require("tsify");
 
 gulp.task('deploy', () => {
   return gulp.src('build/**/*')
@@ -12,4 +14,17 @@ gulp.task('copy', function() {
         .pipe(gulp.dest('build'));
 });
 
-gulp.task('default', ['copy','deploy']);
+gulp.task("ts7", function () {
+  return browserify({
+    basedir: '7-Logical-Agents',
+    debug: true,
+    entries: ['main.ts'],
+    cache: {},
+    packageCache: {}
+  }).plugin(tsify)
+    .bundle()
+    .pipe(source('bundle.js'))
+    .pipe(gulp.dest("7-Logical-Agents"));
+});
+
+gulp.task('default', ['ts7', 'copy']);
