+++ package.json
@@ -23,11 +23,16 @@
     "babel": "^6.5.2",
     "babel-core": "^6.7.6",
     "babel-register": "^6.7.2",
+    "browserify": "^16.2.3",
     "gulp": "^3.9.1",
     "gulp-gh-pages": "^0.5.4",
+    "gulp-typescript": "^5.0.0",
     "gulp-watch": "^4.3.5",
     "jspm": "^0.16.31",
-    "prettier": "^1.7.0"
+    "prettier": "^1.7.0",
+    "tsify": "^4.0.1",
+    "typescript": "^3.2.2",
+    "vinyl-source-stream": "^2.0.0"
   },
   "jspm": {
     "dependencies": {
